refactor : Lotto에 해당하는 상수값을 AutoLottoGenerator 객체에서 Lotto객체로 이동
